requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
PyJWT==1.7.1
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
